changed text preprpocesing from doug stage.
first, cleaned up the code.
secodn, wrote it to preserve original approach
third, changed the merging of lags and leads to handle na's differently
(to tolerate them more).  this increases the output data and decrases by
10% the accuracy of a the classifier.  i don't understand it actually.
fourth, added features that were intended to imprvoe the integration of
bert features with existing features.  the point was that fully
leveraging bert meant zipping two different tokenizers, which was hard.